[libra]REC: diff logic x -> new engine/diff-engine module
Created a new engine module to handle some of the shared mechanisms between app layers
Reconstructed the diff logic inside libra/commands/diff.rs, and moved the new logic to the diff_engine
