+++ Cargo.toml
@@ -18,6 +18,7 @@ members = [
     "extensions/observatory",
     "scorpio",
     "context",
+    "engine",
 ]
 default-members = ["mega", "mono", "libra", "aries", "orion", "orion-server"]
 resolver = "1"
@@ -36,6 +37,7 @@ mono = { path = "mono" }
 observatory = { path = "extensions/observatory" }
 orion = { path = "orion" }
 context = { path = "context" }
+engine = { path = "engine" }
 
 anyhow = "1.0.98"
 serde = "1.0.219"
