+++ engine/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod diff_engine;
+
+pub use diff_engine::DiffEngine;
