+++ libra/Cargo.toml
@@ -17,6 +17,7 @@ ceres = { workspace = true }
 clap = { workspace = true, features = ["derive"] }
 colored = { workspace = true }
 common = { workspace = true }
+engine = { workspace = true }
 flate2 = { workspace = true } # add features = ["zlib"] if slow
 futures = { workspace = true }
 futures-util = { workspace = true }
