+++ libra/src/command/diff.rs
@@ -1,12 +1,10 @@
 use std::{
-    collections::{HashMap, HashSet},
     fmt,
     io::{self, Write},
-    path::{Path, PathBuf},
+    path::{PathBuf},
 };
 
 use clap::Parser;
-use imara_diff::{Algorithm, BasicLineDiffPrinter, Diff, InternedInput, UnifiedDiffConfig};
 use mercury::{
     hash::SHA1,
     internal::{
@@ -15,6 +13,7 @@ use mercury::{
         pack::utils::calculate_object_hash,
     },
 };
+use engine::DiffEngine;
 use similar;
 
 use crate::{
@@ -28,8 +27,7 @@ use crate::{
 
 #[cfg(unix)]
 use std::process::{Command, Stdio};
-
-use crate::utils::path_ext::PathExt;
+use crate::utils::util::{to_workdir_path};
 
 #[derive(Parser, Debug)]
 pub struct DiffArgs {
@@ -125,13 +123,29 @@ pub async fn execute(args: DiffArgs) {
     let paths: Vec<PathBuf> = args.pathspec.iter().map(util::to_workdir_path).collect();
 
     let mut buf: Vec<u8> = Vec::new();
+    let read_content = |file: &PathBuf, hash: &SHA1| {
+        // read content from blob or file
+        match load_object::<Blob>(hash) {
+            Ok(blob) => blob.data,
+            Err(_) => {
+                let file = to_workdir_path(file);
+                std::fs::read(&file)
+                    .map_err(|e| {
+                        eprintln!("fatal: could not read file '{}': {}", file.display(), e);
+                    })
+                    .unwrap()
+            }
+        }
+    };
+
     // filter files, cross old and new files, and pathspec
-    diff(
+    DiffEngine::diff(
         old_blobs,
         new_blobs,
         args.algorithm.unwrap_or_default(),
         paths.into_iter().collect(),
         &mut buf,
+        &read_content,
     )
     .await;
 
@@ -161,123 +175,6 @@ pub async fn execute(args: DiffArgs) {
     }
 }
 
-pub async fn diff(
-    old_blobs: Vec<(PathBuf, SHA1)>,
-    new_blobs: Vec<(PathBuf, SHA1)>,
-    algorithm: String,
-    filter: Vec<PathBuf>,
-    w: &mut dyn io::Write,
-) {
-    let old_blobs: HashMap<PathBuf, SHA1> = old_blobs.into_iter().collect();
-    let new_blobs: HashMap<PathBuf, SHA1> = new_blobs.into_iter().collect();
-    // unison set
-    let union_files: HashSet<PathBuf> = old_blobs.keys().chain(new_blobs.keys()).cloned().collect();
-    tracing::debug!(
-        "old blobs {:?}, new blobs {:?}, union files {:?}",
-        old_blobs.len(),
-        new_blobs.len(),
-        union_files.len()
-    );
-
-    let read_content = |file: &PathBuf, hash: &SHA1| {
-        // read content from blob or file
-        match load_object::<Blob>(hash) {
-            Ok(blob) => blob.data,
-            Err(_) => {
-                let file = util::workdir_to_absolute(file);
-                std::fs::read(&file)
-                    .map_err(|e| {
-                        eprintln!("fatal: could not read file '{}': {}", file.display(), e);
-                    })
-                    .unwrap()
-            }
-        }
-    };
-
-    // filter files, cross old and new files, and pathspec
-    for file in union_files {
-        // if new_file did't start with any path in filter, skip it
-        if !filter.is_empty() && !filter.iter().any(|path| file.sub_of(path)) {
-            continue;
-        }
-
-        let new_hash = new_blobs.get(&file);
-        let old_hash = old_blobs.get(&file);
-        if new_hash == old_hash {
-            continue;
-        }
-
-        let old_content = match old_hash.as_ref() {
-            Some(hash) => read_content(&file, hash),
-            None => Vec::new(),
-        };
-        let new_content = match new_hash.as_ref() {
-            Some(hash) => read_content(&file, hash),
-            None => Vec::new(),
-        };
-
-        writeln!(
-            w,
-            "diff --git a/{} b/{}",
-            file.display(),
-            file.display() // files name is always the same, current did't support rename
-        )
-        .unwrap();
-
-        if old_hash.is_none() {
-            writeln!(w, "new file mode 100644").unwrap();
-        } else if new_hash.is_none() {
-            writeln!(w, "deleted file mode 100644").unwrap();
-        }
-
-        let old_index = old_hash.map_or("0000000".to_string(), |h| h.to_string()[0..8].to_string());
-        let new_index = new_hash.map_or("0000000".to_string(), |h| h.to_string()[0..8].to_string());
-        writeln!(w, "index {old_index}..{new_index}").unwrap();
-        // check is the content is valid utf-8 or maybe binary
-        let old_type = infer::get(&old_content);
-        let new_type = infer::get(&new_content);
-        match (
-            String::from_utf8(old_content),
-            String::from_utf8(new_content),
-        ) {
-            (Ok(old_text), Ok(new_text)) => {
-                let (old_prefix, new_prefix) = if old_text.is_empty() {
-                    // New file
-                    (
-                        "/dev/null".to_string(),
-                        format!("b/{}", file_display(&file, new_hash, new_type)),
-                    )
-                } else if new_text.is_empty() {
-                    // Remove file
-                    (
-                        format!("a/{}", file_display(&file, old_hash, old_type)),
-                        "/dev/null".to_string(),
-                    )
-                } else {
-                    // Update file
-                    (
-                        format!("a/{}", file_display(&file, old_hash, old_type)),
-                        format!("b/{}", file_display(&file, new_hash, new_type)),
-                    )
-                };
-                writeln!(w, "--- {old_prefix}").unwrap();
-                writeln!(w, "+++ {new_prefix}").unwrap();
-                imara_diff_result(&old_text, &new_text, algorithm.as_str(), w);
-            }
-            _ => {
-                // TODO: Handle non-UTF-8 data as binary for now; consider optimization in the future.
-                writeln!(
-                    w,
-                    "Binary files a/{} and b/{} differ",
-                    file_display(&file, old_hash, old_type),
-                    file_display(&file, new_hash, new_type)
-                )
-                .unwrap();
-            }
-        }
-    }
-}
-
 async fn get_commit_blobs(commit_hash: &SHA1) -> Vec<(PathBuf, SHA1)> {
     let commit = load_object::<Commit>(commit_hash).unwrap();
     let tree = load_object::<Tree>(&commit.tree_id).unwrap();
@@ -299,25 +196,6 @@ fn get_files_blobs(files: &[PathBuf]) -> Vec<(PathBuf, SHA1)> {
             (p.to_owned(), calculate_object_hash(ObjectType::Blob, &data))
         })
         .collect()
-}
-
-// display file with type
-fn file_display(file: &Path, hash: Option<&SHA1>, file_type: Option<infer::Type>) -> String {
-    let file_name = match hash {
-        Some(_) => file.display().to_string(),
-        None => "dev/null".to_string(),
-    };
-
-    if let Some(file_type) = file_type {
-        // Check if the file type is displayable in browser, like image, audio, video, etc.
-        if matches!(
-            file_type.matcher_type(),
-            infer::MatcherType::Audio | infer::MatcherType::Video | infer::MatcherType::Image
-        ) {
-            return format!("{} ({})", file_name, file_type.mime_type()).to_string();
-        }
-    }
-    file_name
 }
 
 struct Line(Option<usize>);
@@ -362,39 +240,11 @@ fn similar_diff_result(old: &str, new: &str, w: &mut dyn io::Write) {
     }
 }
 
-fn imara_diff_result(old: &str, new: &str, algorithm: &str, w: &mut dyn io::Write) {
-    let input = InternedInput::new(old, new);
-
-    let algo = match algorithm {
-        "myers" => Algorithm::Myers,
-        "myersMinimal" => Algorithm::MyersMinimal,
-        // default is the histogram algo
-        _ => Algorithm::Histogram,
-    };
-    tracing::debug!("libra [diff]: choose the algorithm: {:?}", algo);
-
-    let mut diff = Diff::compute(algo, &input);
-
-    // did the postprocess_lines
-    diff.postprocess_lines(&input);
-
-    let result = diff
-        .unified_diff(
-            &BasicLineDiffPrinter(&input.interner),
-            UnifiedDiffConfig::default(),
-            &input,
-        )
-        .to_string();
-
-    write!(w, "{result}").unwrap();
-}
-
 #[cfg(test)]
 mod test {
     use crate::utils::test;
     use serial_test::serial;
     use std::fs;
-    use std::time::Instant;
     use tempfile::tempdir;
 
     use super::*;
@@ -486,71 +336,5 @@ mod test {
         let blob = get_files_blobs(&[PathBuf::from("should_ignore"), PathBuf::from("not_ignore")]);
         assert_eq!(blob.len(), 1);
         assert_eq!(blob[0].0, PathBuf::from("not_ignore"));
-    }
-
-    #[test]
-    fn test_diff_algorithms_correctness_and_efficiency() {
-        let old = r#"function foo() {
-    if (condition) {
-        doSomething();
-        doSomethingElse();
-        andAnotherThing();
-    } else {
-        alternative();
-    }
-}"#;
-
-        let new = r#"function foo() {
-    if (condition) {
-        // Added comment
-        doSomething();
-        // Modified this line
-        modifiedSomethingElse();
-        andAnotherThing();
-    } else {
-        alternative();
-    }
-
-    // Added new block
-    addedNewFunctionality();
-}"#;
-        let mut outputs = Vec::new();
-
-        let algos = ["histogram", "myers", "myersMinimal"];
-
-        // test the different algo benchmark
-        for algo in algos {
-            let mut buf = Vec::new();
-            let start = Instant::now();
-            imara_diff_result(old, new, algo, &mut buf);
-            let elapse = start.elapsed();
-            let ouput = String::from_utf8(buf).expect("Invalid UTF-8 in diff ouput");
-
-            println!("libra diff algorithm: {algo:?} Spend Time: {elapse:?}");
-            assert!(
-                !ouput.is_empty(),
-                "libra diff algorithm: {algo} produce a empty output"
-            );
-            assert!(
-                ouput.contains("@@"),
-                "libra diff algorithm: {algo}, ouput missing diff markers"
-            );
-
-            outputs.push((algo, ouput));
-        }
-
-        // check the line counter difference
-        for (algo, output) in outputs {
-            let plus_line = output.lines().filter(|line| line.starts_with("+")).count();
-            let minus_line = output.lines().filter(|line| line.starts_with("-")).count();
-            assert_eq!(
-                plus_line, 6,
-                "libra diff algorithm {algo}, expect plus_line: 6, got {plus_line} "
-            );
-            assert_eq!(
-                minus_line, 1,
-                "libra diff algorithm {algo}, expect minus_line: 1, got {minus_line} "
-            );
-        }
     }
 }
